Split of PreferencesViewController in separate files (in progress) Add the new MacMini and MacBook Airs as supporting HEVC Main10 (oops !)
